Extract hashers as a separate file
